Fix naive and two_step_slow dct

+++ include/dct/loeffler_float.h
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
 
-void dct_loeffler_float(const uint8_t data_in[8][8], uint8_t data_out[8][8]);
+void dct_loeffler_float(const uint8_t data_in[8][8], int16_t data_out[8][8]);
 
 #endif // SENG440_LOEFFLER_FLOAT_H
